fix: set GH_REPO env var in Go license script
I tried running this script for the first time this morning and it failed. Looking into how this works in the Copywrite tool led me to https://github.com/cli/go-gh#go-library-for-the-github-cli which got me hoping that maybe all this is missing is this `GH_REPO` env var. Might be a long shot, though.