feat: automate upgrading CDKTF (#222)
This workflow adapts scripts we're using elsewhere already like
https://github.com/hashicorp/cdktf-aws-cdk/blob/main/.github/workflows/upgrade-cdktf.yml
and
https://github.com/hashicorp/terraform-cdk-action/blob/main/.github/workflows/upgrade-cdktf.yml
to automatically create a PR to upgrade CDKTF when new (non-patch)
versions come out. This variant creates two PRs:

- one to upgrade CDKTF on the prebuilt provider repos; for now, I've
left the automerge label off so it will depend on us to manually approve
the PR before it is merged (if anyone disagrees with that, let me know,
but I feel like that's generally how we've done things)
- a draft PR to upgrade CDKTF on this repo, which will depend on the
above PR being merged first (because the GitHub provider will need to be
updated before this repo can do the upgrade)

Hopefully this will help prevent situations in the future like we had
with this repo not updating its version of CDKTF for 6+ months.