chore(ci): temporarily hardcode jsii as a non-breaking version change
Because of the earlier botched rollout, the collect-changes script
thinks JSII is being upgraded from v0.6.0 to v5.4.0 and thus treats
this as a major version change for all prebuilt providers. In reality,
though, nothing is actually changing since the version of JSII is
already set to v5.4.0.

The easiest fix is to just have the script temporarily treat any/all
JSII version changes as non-breaking. This change should be reverted
as soon as the next deploy is successful.