chore: upgrade self to cdktf 0.21.0 (#393)
This PR upgrades CDKTF to version `0.21.0` in `cdktf-repository-manager`
(this repo).
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:

- [x] Wait for #392 to be merged and for the deployment to succeed
- [x] Upgrade `@cdktf/provider-github` to a version compatible with
`cdktf@0.21.0` (`yarn add ...`)

Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Mutahhir Hayat <mutahhir.hayat@hashicorp.com>
Co-authored-by: Nara Kasbergen <nara.kasbergen@gmail.com>