feat: deprecate and archive prebuilt bindings for google provider (#470)
HashiCorp has made the decision to stop publishing new versions of
prebuilt Terraform `google` provider bindings for
[CDK for Terraform](https://cdk.tf). Once this PR is merged, the
[cdktf/cdktf-provider-google](https://github.com/cdktf/cdktf-provider-google)
repository will be archived and will no longer
be supported in any way by HashiCorp. Previously-published versions of
the prebuilt `google` provider will still continue
to be available on their respective package managers (e.g. npm, PyPi,
Maven, NuGet), but these will not be compatible with
new releases of `cdktf` and are no longer eligible for commercial
support.

Please complete the following steps in this exact order to complete the
deprecation process:

- [x] Double-check `provider.json`, `sharded-stacks.json`, and
`providersWithCustomRunners.json` in this PR for any syntax errors
caused by extraneous commas
- [x] Mark this PR as ready for review and examine the plan output from
the checks to confirm the correct resources are destroyed
- [x] Approve and merge cdktf/cdktf-provider-google#2087 and ensure that
the release is published to all package managers
- N.B. New published versions take 6~8 hours to show up in Maven, but
you do not need to wait for that, just ensure that the `release_maven`
job completed successfully
- [ ] [Manually
deprecate](https://learn.microsoft.com/en-us/nuget/nuget-org/deprecate-packages)
the `HashiCorp.Cdktf.Providers.Google` package in NuGet Gallery
_(optional but recommended)_
- Provide the following custom message: _HashiCorp is no longer
publishing new versions of the prebuilt provider for google.
Previously-published versions of this prebuilt provider will still
continue to be available as installable packages on NuGet, but these
will not be compatible with newer versions of CDK for Terraform and are
not eligible for commercial support. You can continue to use the google
provider in your CDK for Terraform projects with newer versions of
CDKTF, but you will need to generate the bindings locally. See
https://cdk.tf/imports for details._
- [x] Remove the "do-not-merge" label and merge this PR

Please also ensure that not too much time passes in between each of
these steps. Notably, if the PR in the provider repo is
merged but other changes are deployed before that repo is properly
archived, there could be unintended behavior. So, it is
highly recommended that you complete the above steps in short
succession.

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen Kwon <855115+xiehan@users.noreply.github.com>